make MMX & SSE work on amd64 :)
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@5789 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/video/FBPostProcessor.hh
@@ -37,7 +37,7 @@ private:
 	void preCalcNoise(double factor);
 	void drawNoise();
 	void drawNoiseLine(Pixel* in, Pixel* out, signed char* noise,
-	                   unsigned width);
+	                   unsigned long width);
 
 	// Observer<Setting>
 	virtual void update(const Setting& setting);
