+++ src/video/Scanline.hh
@@ -57,7 +57,7 @@ public:
 	  * @param width Line width in pixels.
 	  */
 	void draw(const Pixel* src1, const Pixel* src2, Pixel* dst,
-	          unsigned factor, unsigned width);
+	          unsigned factor, unsigned long width);
 
 	/** Darken one pixel. Typically used to implement drawBlank().
 	 */
